requirements.txt
# Core trading dependencies
alpaca-py==0.42.1
python-dotenv==1.1.1
pandas==2.3.2
numpy==2.3.3
requests==2.32.5

# Machine Learning dependencies
scikit-learn==1.7.2
xgboost==2.1.3
lightgbm==4.5.0
imbalanced-learn==0.12.3

# Technical Analysis
TA-Lib==0.4.28

# Data processing
scipy==1.16.2
matplotlib==3.10.6
seaborn==0.13.2

# News and sentiment
finnhub-python==2.4.18

# Utilities
schedule==1.2.2
websockets==15.0.1
joblib==1.5.2

# Development and monitoring
jupyter==1.1.1
notebook==7.3.1
